+++ Editor/Interfaces/IAudioClipModuleAnalyzer.cs
@@ -1,0 +1,13 @@
+using System.Collections.Generic;
+using Unity.ProjectAuditor.Editor;
+using Unity.ProjectAuditor.Editor.Core;
+using UnityEditor;
+using UnityEngine;
+
+namespace Unity.ProjectAuditor.Editor.Interfaces
+{
+    internal interface IAudioClipModuleAnalyzer : IModuleAnalyzer
+    {
+        IEnumerable<ProjectIssue> Analyze(ProjectAuditorParams projectAuditorParams, AudioImporter audioImporter);
+    }
+}